Inline projects tests, split calendar modules, drop section markers
Merge the projects test module back into projects.rs as an inline
cfg(test) module, which is idiomatic Rust; the plan now states that
tests stay inline.
Split calendar/events.rs into events/classify.rs, resolve.rs, order.rs,
and absences.rs, and calendar/caldav.rs into caldav/report.rs and
caldav/write.rs, each with co-located tests.
Remove all remaining section marker comments, including the backlog-ID
markers in the frontend spec files, and drop narration comments in the
touched files.
