+++ src-tauri/src/integrations/calendar/caldav/mod.rs
@@ -1,0 +1,8 @@
+mod report;
+mod write;
+
+pub(super) use report::fetch_calendar_events;
+pub(crate) use write::{
+    create_assignment_core, delete_assignment_core, update_assignment_core, AssignmentWrite,
+    CaldavSession,
+};
