+++ src-tauri/src/integrations/calendar/events/mod.rs
@@ -1,0 +1,9 @@
+mod absences;
+mod classify;
+mod order;
+mod resolve;
+
+pub(super) use absences::map_absence_raw_events_for_week;
+pub(super) use classify::classify_event;
+pub(super) use order::sort_events_absences_first;
+pub(super) use resolve::resolve_event;
