+++ src/app/components/assignment-modal.spec.tsx
@@ -133,7 +133,6 @@ describe("AssignmentModal", () => {
   });
 });
 
-// ── 6.1 – result list display vs empty default state ──────────────────────────
 describe("ProjectResultList", () => {
   it("renders the filtered projects as selectable options", () => {
     const html = renderToStaticMarkup(
@@ -165,7 +164,6 @@ describe("ProjectResultList", () => {
     expect(html).toBe("");
   });
 
-  // ── 6.3 – keyboard selection highlight ──────────────────────────────────────
   it("marks the highlighted option as selected", () => {
     const html = renderToStaticMarkup(
       <ProjectResultList
@@ -186,7 +184,6 @@ describe("ProjectResultList", () => {
   });
 });
 
-// ── 6.3 – arrow key navigation over the displayed list ────────────────────────
 describe("nextHighlightIndex", () => {
   it("moves down from the unhighlighted state to the first item", () => {
     expect(nextHighlightIndex(-1, 3, 1)).toBe(0);
@@ -207,7 +204,6 @@ describe("nextHighlightIndex", () => {
   });
 });
 
-// ── BL-031 5.1 / 5.2 / 5.3 – default suggestions in the result list ───────────
 describe("default suggestions rendering", () => {
   const overdue = [
     project("Projekt 10", "/v1/projects/10"),
@@ -265,7 +261,6 @@ describe("default suggestions rendering", () => {
   });
 });
 
-// ── BL-031 5.7 – clearing the filter restores the default suggestions ─────────
 describe("resolveDisplayedProjects", () => {
   const suggestions = [project("Projekt Zuletzt", "/v1/projects/99")];
   const results = [project("Projekt Nord", "/v1/projects/10")];
@@ -293,7 +288,6 @@ describe("resolveDisplayedProjects", () => {
   });
 });
 
-// ── BL-031 5.6 – empty state message display ──────────────────────────────────
 describe("SuggestionEmptyState", () => {
   it("shows the German message when no suggestions are available", () => {
     const html = renderToStaticMarkup(
@@ -336,7 +330,6 @@ describe("SuggestionEmptyState", () => {
   });
 });
 
-// ── 6.5 – Escape precedence (clear vs close) ──────────────────────────────────
 describe("resolveEscapeAction", () => {
   it("clears a non-empty filter instead of closing", () => {
     expect(resolveEscapeAction("Nord")).toBe("clear");
@@ -347,7 +340,6 @@ describe("resolveEscapeAction", () => {
   });
 });
 
-// ── bl-033 1.1 / 1.3 – only a create carries a next-day ghost payload ─────────
 describe("resolveSaveAction", () => {
   it("builds a create action carrying the saved project", () => {
     expect(
